requirements.txt
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
Flask-Login>=0.6.3
python-dateutil>=2.9.0
APScheduler>=3.10.4
pytz>=2024.1
cryptography>=42.0.0
Jinja2>=3.1.4
click>=8.1.7
itsdangerous>=2.2.0
MarkupSafe>=2.1.5
SQLAlchemy>=2.0.31
blinker>=1.8.2
flask-migrate
apscheduler
python-dateutil
pytz